--- tests/test_self_use.py
@@ -1,59 +1,0 @@
-from collections.abc import Callable
-from typing import Generic
-
-from guppylang.decorator import guppy
-from guppylang.std import quantum as phys
-from guppylang.std.builtins import array, comptime, owned
-from hugr.package import Package
-
-import qcorrect as qct
-
-N = guppy.nat_var("N")
-
-
-@qct.block
-class CodeBlock(Generic[N]):
-    data_qs: array[phys.qubit, N]
-
-
-class CodeDef(qct.CodeDefinition):
-    n: int
-
-    @qct.operation(op=phys.qubit)
-    def zero(self) -> Callable:
-        @guppy
-        def circuit() -> "CodeBlock[comptime(self.n)]":
-            return CodeBlock(array(phys.qubit() for _ in range(comptime(self.n))))
-
-        return circuit
-
-    @qct.operation(op=phys.qubit)
-    def zero_call(self) -> Callable:
-        @guppy
-        def circuit() -> "CodeBlock[comptime(self.n)]":
-            return self.zero()
-
-        return circuit
-
-    @qct.operation(op=phys.measure)
-    def measure(self) -> Callable:
-        @guppy
-        def circuit(
-            q: "CodeBlock[comptime(self.n)] @ owned",
-        ) -> "array[bool, comptime(self.n)]":
-            return phys.measure_array(q.data_qs)
-
-        return circuit
-
-
-def test_self_use():
-    code = CodeDef(n=5)
-
-    @guppy
-    def main() -> None:
-        q = code.zero_call()
-        code.measure(q)
-
-    hugr = main.compile()
-
-    assert isinstance(hugr, Package)
